flask_backend/requirements.txt
flask
flask-cors
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
gunicorn
openai